test(gateway): split provider bootstrap coverage
